custom_components/loxone/pyloxone_api/requirements.txt
aiohttp
websockets
httpx
bottle
pycryptodome